+++ changelog.md
@@ -3,5 +3,7 @@
 ## General
 
 ## lace
+- Add new `StructInitExpr` for struct initializer expressions.
 
 ## LIR
+- Updated SSA rewrite pass to new LIR definitions.